+++ src/main/java/mdteam/ait/core/blocks/CoralPlantBlock.java
@@ -103,7 +103,7 @@ public class CoralPlantBlock extends HorizontalDirectionalBlock implements Block
     private void createTardis(ServerWorld world, BlockPos pos) {
         // Create a new tardis
         ServerTardis created = ServerTardisManager.getInstance().create(new AbsoluteBlockPos.Directed(pos, world, Direction.NORTH), CategoryRegistry.getInstance().get(GrowthCategory.REFERENCE), ExteriorVariantRegistry.getInstance().get(CoralGrowthVariant.REFERENCE), DesktopRegistry.DEFAULT_CAVE, false);
-        created.getHandlers().getFuel().setFuel(0);
+        created.getHandlers().getFuel().setCurrentFuel(0);
         // created.getHandlers().getOvergrown().setOvergrown(true); //fixme created.getEnvironmentHandler().setCoralCovered(true);
 
         //LightningEntity lightning = new LightningEntity(EntityType.LIGHTNING_BOLT, world);