+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Max increment changed to 10,000.
 - Max fuel amount changed to 50,000au.
 - Fixed telepathic circuit being broken
+- Added functionality to the hammer, hitting the console with it now speeds up flight at a cost 😉